Shuffle things that need to be integration tests into `telemetry_reconciler_test.go`. Implement some basic unit tests for Resourcer and implementations.
