+++ controllers/k8ssandra/telemetry_reconciler_test.go
@@ -1,3 +1,33 @@
 package k8ssandra
 
-//TODO
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_isPrometheusInstalled(t *testing.T) {
+	assert.Fail(t, "not implemented")
+}
+
+// TestCassPrometheusResourcer_UpdateResources_SUCCESS tests that NewServiceMonitor succeeds in creating a ServiceMonitor.
+func TestCassPrometheusResourcer_UpdateResources_SUCCESS(t *testing.T) {
+	assert.Fail(t, "not implemented")
+}
+
+// TestCassPrometheusResourcer_CleanupResources_FAIL_Incomplete tests that the correct error type is returned when an incomplete
+// CassandraSMConfig is passed to NewServiceMonitor.
+func TestCassPrometheusResourcer_CleanupResources_FAIL_Incomplete(t *testing.T) {
+	assert.Fail(t, "not implemented")
+}
+
+// TestCassTelemetryResourcer_CreateResources_SUCCESS tests that NewServiceMonitor succeeds in creating the expected resources.
+func TestCassTelemetryResourcer_CreateResources_SUCCESS(t *testing.T) {
+	assert.Fail(t, "not implemented")
+}
+
+// TestCassTelemetryResourcer_CleanupResources_FAIL_Incomplete tests that the correct error type is returned when an incomplete
+// CassandraSMConfig is passed to NewServiceMonitor.
+func TestCassTelemetryResourcer_CleanupResources_FAIL_Incomplete(t *testing.T) {
+	assert.Fail(t, "not implemented")
+}
