+++ pkg/telemetry/cass_telemetry_resourcer_test.go
@@ -8,18 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-// TestCassTelemetryResourcer_CreateResources_SUCCESS tests that NewServiceMonitor succeeds in creating the expected resources.
-func TestCassTelemetryResourcer_CreateResources_SUCCESS(t *testing.T) {
-	assert.Fail(t, "not implemented")
-}
-
-// TestCassTelemetryResourcer_CleanupResources_FAIL_Incomplete tests that the correct error type is returned when an incomplete
-// CassandraSMConfig is passed to NewServiceMonitor.
-func TestCassTelemetryResourcer_CleanupResources_FAIL_Incomplete(t *testing.T) {
-	assert.Fail(t, "not implemented")
-}
-
 // TestCassTelemetryResourcer_IS_TelemetryResourcer tests that CassTelemetryResourcer implements the Resourcer interface.
 func TestCassTelemetryResourcer_IS_TelemetryResourcer(t *testing.T) {
-	assert.Fail(t, "This should be implemented in an envtest probably")
+	assert.Implements(t, (*Resourcer)(nil), CassTelemetryResourcer{})
 }
